refactor: Split monolithic apple_mail_mcp.py into modular package
Splits the 3,802-line single-file MCP server into a well-organized
package with dedicated modules while preserving all 26 tool signatures,
return formats, and backward compatibility.

New structure:
- apple_mail_mcp/server.py: FastMCP instance + user preferences
- apple_mail_mcp/core.py: Shared helpers + AppleScript template functions
- apple_mail_mcp/constants.py: Newsletter patterns, skip folders, etc.
- apple_mail_mcp/tools/inbox.py: 6 inbox tools
- apple_mail_mcp/tools/search.py: 8 search tools
- apple_mail_mcp/tools/compose.py: 4 composition tools
- apple_mail_mcp/tools/manage.py: 4 management tools
- apple_mail_mcp/tools/analytics.py: 4 analytics tools + helper

Also fixes search_by_sender and get_recent_from_sender timeouts by
defaulting to INBOX-only search and using direct account access.
