+++ apple_mail_mcp/server.py
@@ -1,0 +1,10 @@
+"""FastMCP server instance and user preferences."""
+
+import os
+from mcp.server.fastmcp import FastMCP
+
+# Initialize FastMCP server
+mcp = FastMCP("Apple Mail MCP")
+
+# Load user preferences from environment
+USER_PREFERENCES = os.environ.get("USER_EMAIL_PREFERENCES", "")
